+++ examples/table_basic.rs
@@ -27,7 +27,6 @@ use std::io::{stdout, Stdout};
 use std::time::{Duration, SystemTime};
 
 mod data;
-mod table_insane_offset;
 
 fn main() -> Result<(), anyhow::Error> {
     setup_logging()?;
@@ -229,14 +228,6 @@ fn repaint_table(frame: &mut Frame<'_>, area: Rect, data: &mut Data, state: &mut
             self.0.len()
         }
 
-        fn row_height(&self, _row: usize) -> u16 {
-            1
-        }
-
-        fn row_style(&self, _r: usize) -> Style {
-            Style::default()
-        }
-
         fn render_cell(&self, column: usize, row: usize, area: Rect, buf: &mut Buffer) {
             if let Some(d) = self.0.get(row) {
                 match column {
@@ -290,11 +281,11 @@ fn repaint_table(frame: &mut Frame<'_>, area: Rect, data: &mut Data, state: &mut
                 Cell::from("Val2"),
                 Cell::from("State"),
             ])
-            .style(Style::new().black().bg(Color::from_u32(0x98c379))),
+            .style(Some(Style::new().black().bg(Color::from_u32(0x98c379)))),
         )
         .footer(
             Row::new(["a", "b", "c", "d", "e"])
-                .style(Style::new().black().bg(Color::from_u32(0x98c379))),
+                .style(Some(Style::new().black().bg(Color::from_u32(0x98c379)))),
         )
         .flex(Flex::End)
         .style(Style::default().bg(Color::Rgb(25, 25, 25)));