+++ examples/table_layoutwidth.rs
@@ -230,14 +230,6 @@ fn repaint_table(frame: &mut Frame<'_>, area: Rect, data: &mut Data, state: &mut
             self.0.len()
         }
 
-        fn row_height(&self, _row: usize) -> u16 {
-            1
-        }
-
-        fn row_style(&self, _r: usize) -> Style {
-            Style::default()
-        }
-
         fn render_cell(&self, column: usize, row: usize, area: Rect, buf: &mut Buffer) {
             if let Some(d) = self.0.get(row) {
                 match column {
@@ -292,11 +284,11 @@ fn repaint_table(frame: &mut Frame<'_>, area: Rect, data: &mut Data, state: &mut
                 Cell::from("Val2"),
                 Cell::from("State"),
             ])
-            .style(Style::new().black().bg(Color::from_u32(0x98c379))),
+            .style(Some(Style::new().black().bg(Color::from_u32(0x98c379)))),
         )
         .footer(
             Row::new(["a", "b", "c", "d", "e"])
-                .style(Style::new().black().bg(Color::from_u32(0x98c379))),
+                .style(Some(Style::new().black().bg(Color::from_u32(0x98c379)))),
         )
         .flex(Flex::SpaceBetween)
         .style(Style::default().bg(Color::Rgb(25, 25, 25)));